+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ### Changed
 * `multidispatch` supports variable keyword arguments
 * `multidispatch` base implementation always matches
-* Type aliases supported
+* Type and Annotated aliases supported
 
 ## [1.11.2](https://pypi.org/project/multimethod/1.11.2/) - 2024-02-27
 ### Fixed